+++ lib/src/defs.dart
@@ -1,9 +1,0 @@
-library angel_framework.defs;
-import '../common.dart';
-export '../common.dart';
-
-/// Use [Model] instead.
-@deprecated
-class MemoryModel {
-  int id;
-}