+++ lib/src/fast_name_from_symbol.dart
@@ -1,6 +1,10 @@
+final Map<Symbol, String> _cache = {};
+
 String fastNameFromSymbol(Symbol s) {
-  String str = s.toString();
-  int open = str.indexOf('"');
-  int close = str.lastIndexOf('"');
-  return str.substring(open + 1, close);
+  return _cache.putIfAbsent(s, () {
+    String str = s.toString();
+    int open = str.indexOf('"');
+    int close = str.lastIndexOf('"');
+    return str.substring(open + 1, close);
+  });
 }