+++ lib/src/http/base_middleware.dart
@@ -7,8 +7,3 @@ import 'response_context.dart';
 abstract class AngelMiddleware {
   Future<bool> call(RequestContext req, ResponseContext res);
 }
-
-@Deprecated('Use AngelMiddleware instead')
-abstract class BaseMiddleware {
-  Future<bool> call(RequestContext req, ResponseContext res);
-}