+++ lib/src/http/fatal_error.dart
@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:io';
 
 /// Thrown whenever Angel completely fails to respond to a request.
@@ -5,7 +6,8 @@ class AngelFatalError {
   var error;
   HttpRequest request;
   StackTrace stack;
+  Zone zone;
 
-  AngelFatalError({this.request, this.error, this.stack});
+  AngelFatalError({this.request, this.error, this.stack, this.zone});
 }